Workaround issue 115: allow 26 or 27 acoustic events. Results vary depending on some library versions. Probably due to floating point numerical error. See https://gitlab.com/yade-dev/trunk/issues/115 for the list of suspects.

+++ scripts/checks-and-tests/checks/checkJCFpm.py
@@ -47,9 +47,9 @@ AEdata = np.loadtxt('moments_.txt',skiprows=1)
 cracksdata = np.loadtxt('cracks_.txt',skiprows=1)
 
 if not len(cracksdata)==sum(AEdata[:,5]):  # number of cracks should be equivalent to sum of cluster counts
-	raise YadeCheckError('JCFpm checktest: cracks or AE clustering algorithm incorrect')
-if not len(AEdata)==26:
-	raise YadeCheckError('JCFpm checktest: number of acoustic emission events incorrect')
+	raise YadeCheckError('JCFpm checktest: cracks or AE clustering algorithm incorrect: len(cracksdata)=' + str(len(cracksdata)) + ', sum(AEdata[:,5])='+str(sum(AEdata[:,5])))
+if not ((len(AEdata)==26) or (len(AEdata)==27)):
+	raise YadeCheckError('JCFpm checktest: number of acoustic emission events incorrect = ' + str(len(AEdata)))
 
 os.remove('moments_.txt')
 os.remove('cracks_.txt')
